Fix out-of-order texture encode bug [WIP]
Fixes a bug where if the graphics thread needs to duplicate frames, it
would still try to cycle the next texture before the texture was
finished duplicating.

This also instead duplicates the last item in the queue (after the queue
is started) for better performance so extra copies don't need to occur
unnecessarily.